issue-34 refactor: created dedicated put and get packages + moving classes there
