+++ src/main/scala/org/dynamite/action/get/response.scala
@@ -1,0 +1,13 @@
+package org.dynamite.action.get
+
+import org.dynamite.dsl.JsonDeserializable
+import org.json4s._
+
+private[dynamite] case class GetItemResponse(item: JValue)
+
+private[dynamite] object GetItemResponse {
+  implicit val fromJson = new JsonDeserializable[GetItemResponse] {
+    override def deserialize(jValue: JValue): GetItemResponse =
+      GetItemResponse(jValue \ "Item")
+  }
+}
