+++ src/main/scala/org/dynamite/action/put/result.scala
@@ -1,0 +1,5 @@
+package org.dynamite.action.put
+
+case class GetItemResult[A](item: Option[A])
+
+case class PutItemResult()
