--- src/main/scala/org/dynamite/dsl/results.scala
@@ -1,5 +1,0 @@
-package org.dynamite.dsl
-
-case class GetItemResult[A](item: Option[A])
-
-case class PutItemResult()
